+++ imports/ui/components/MsgType.jsx
@@ -49,6 +49,10 @@ export const MsgType = (props) => {
         return <Badge color="dark"><T>messageTypes.IBCTransfer</T></Badge>;
     case "/cosmos.IBCReceiveMsg":
         return <Badge color="dark"><T>messageTypes.IBCReceive</T></Badge>;
+    case "/agoric.swingset.MsgDeliverInbound":
+        return <Badge color="success"><T>messageTypes.AgoricSend</T></Badge>
+    case "/agoric.swingset.MsgProvision":
+        return <Badge color="success"><T>messageTypes.AgoricProvision</T></Badge>
 
     default:
         return <Badge color="primary">{props.type}</Badge>;