+++ imports/ui/components/Transactions.jsx
@@ -14,6 +14,7 @@ export default class ValidatorTransactions extends Component{
             distributionTxs: {},
             governanceTxs: {},
             slashingTxs: {},
+            IBCTxs: {},
         };  
     }
 
@@ -26,7 +27,8 @@ export default class ValidatorTransactions extends Component{
                     stakingTxs: this.props.stakingTxs,
                     distributionTxs: this.props.distributionTxs,
                     governanceTxs: this.props.governanceTxs,
-                    slashingTxs: this.props.slashingTxs
+                    slashingTxs: this.props.slashingTxs,
+                    IBCTxs: this.props.IBCTxs,
                 })
             }
         }
@@ -43,6 +45,7 @@ export default class ValidatorTransactions extends Component{
                 distributionTxs={this.state.distributionTxs}
                 governanceTxs={this.state.governanceTxs}
                 slashingTxs={this.state.slashingTxs}
+                IBCTxs={this.state.IBCTxs}
             />
         }
         else {