+++ imports/ui/components/TransactionsContainer.js
@@ -64,7 +64,8 @@ export default TransactionsContainer = withTracker((props) => {
         IBCTxs: transactionsExist ? Transactions.find({
             $or: [
                 {"tx.body.messages.@type":"/cosmos.IBCTransferMsg"},
-                {"tx.body.messages.@type":"/cosmos.IBCReceiveMsg"}
+                {"tx.body.messages.@type":"/cosmos.IBCReceiveMsg"},
+                {"tx.body.messages.@type":"/agoric.swingset.MsgDeliverInbound"},
             ]
         }).fetch() : {}
     };